Add aliases to package.json to start backend with npm run server or everything with npm run all
Tired of looking up the command every time I need to start the server so I added more commands to scripts in package.json.
- npm run server to start the backend
- npm run all to start backend and the frontend

Note that the all command depends on having Unix like environment. Works for me on windows though, but that might be because I installed build tools. There may also be side effects to look into because I used a | to run both processes which causes the first process to run and the second process consumes the output, which may cause unintended side effects.